Splitted out the arena config creation into a separate component
