+++ src/RobotSimulator/RobotSimulator.tsx
@@ -3,22 +3,16 @@ import "./RobotSimulator.css";
 import { Sim3D } from "@fruk/simulator-core";
 import { StdWorldBuilder } from "./StdWorldBuilder";
 import { RobotHandle } from "@fruk/simulator-core/dist/engine/handles";
-import { WorldConfig } from "@fruk/simulator-core/dist/engine/SimulatorConfig";
-import {
-  IBallSpec,
-  IBoxSpec,
-  IConeSpec,
-} from "@fruk/simulator-core/dist/engine/specs/CoreSpecs";
 import { useSelector } from "react-redux";
 import { getMotorPower } from "./robotSimulatorSlice";
 import { IVirtualMachine, VMContext } from "../JavascriptVM/JavascriptVM";
-
-interface ArenaConfig {
-  worldConfig: WorldConfig;
-  ballSpecs?: IBallSpec[];
-  boxSpecs?: IBoxSpec[];
-  coneSpecs?: IConeSpec[];
-}
+import {
+  ArenaConfig,
+  setupPlainArena,
+  setupParkingLotArena,
+  setupZigZagArena,
+  setupBowlingArena,
+} from "./ArenaConfigBuilder";
 
 // This component coordinates between react html and the canvas. It uses the 3DSim class to handle the 3D scene and
 // proxies all required events from the browsers into the simulation. All react redux integration is done at this level.
@@ -90,253 +84,26 @@ export const RobotSimulator: FunctionComponent = () => {
       });
     }
 
-    function setupPlainArena() {
-      const arenaConfig: ArenaConfig = {
-        worldConfig: {
-          zLength: 20,
-          xLength: 20,
-          walls: [],
-          camera: {
-            position: {
-              x: 0,
-              y: 8,
-              z: 10,
-            },
-          },
-        },
-      };
-      setupArena(arenaConfig);
-    }
-
-    function setupParkingLotArena() {
-      const height: number = 1;
-      const thickness: number = 0.3;
-      const length: number = 4;
-      const width: number = 3;
-      const arenaConfig: ArenaConfig = {
-        worldConfig: {
-          zLength: 30,
-          xLength: 20,
-          walls: [],
-          camera: {
-            position: {
-              x: 0,
-              y: 8,
-              z: 10,
-            },
-          },
-        },
-        boxSpecs: [
-          {
-            type: "box",
-            dimensions: { x: thickness, y: height, z: length },
-            initialPosition: { x: 9, y: -13 },
-          },
-          {
-            type: "box",
-            dimensions: { x: thickness, y: height, z: length },
-            initialPosition: { x: 5, y: -13 },
-          },
-          {
-            type: "box",
-            dimensions: { x: width, y: height, z: thickness },
-            initialPosition: { x: -8.5, y: -12 },
-          },
-          {
-            type: "box",
-            dimensions: { x: width, y: height, z: thickness },
-            initialPosition: { x: -8.5, y: -8 },
-          },
-        ],
-      };
-      setupArena(arenaConfig);
-    }
-
-    function setupZigZagArena() {
-      const arenaConfig: ArenaConfig = {
-        worldConfig: {
-          zLength: 40,
-          xLength: 20,
-          walls: [],
-          camera: {
-            position: {
-              x: 0,
-              y: 8,
-              z: 10,
-            },
-          },
-        },
-        coneSpecs: [
-          {
-            type: "cone",
-            baseColor: 0x980000,
-            height: 5,
-            radius: 0.5,
-            initialPosition: { x: 0, y: -15 },
-          },
-          {
-            type: "cone",
-            baseColor: 0x980000,
-            height: 5,
-            radius: 0.5,
-            initialPosition: { x: 2, y: -10 },
-          },
-          {
-            type: "cone",
-            baseColor: 0x980000,
-            height: 5,
-            radius: 0.5,
-            initialPosition: { x: -2, y: -5 },
-          },
-          {
-            type: "cone",
-            baseColor: 0x980000,
-            height: 5,
-            radius: 0.5,
-            initialPosition: { x: 2, y: 0 },
-          },
-          {
-            type: "cone",
-            baseColor: 0x980000,
-            height: 5,
-            radius: 0.5,
-            initialPosition: { x: -2, y: 5 },
-          },
-          {
-            type: "cone",
-            baseColor: 0x980000,
-            height: 5,
-            radius: 0.5,
-            initialPosition: { x: 2, y: 10 },
-          },
-          {
-            type: "cone",
-            baseColor: 0x980000,
-            height: 5,
-            radius: 0.5,
-            initialPosition: { x: 0, y: 15 },
-          },
-        ],
-      };
-      setupArena(arenaConfig);
-    }
-
-    function setupBowlingArena() {
-      const arenaConfig: ArenaConfig = {
-        worldConfig: {
-          zLength: 60,
-          xLength: 20,
-          walls: [],
-          camera: {
-            position: {
-              x: 0,
-              y: 10,
-              z: 10,
-            },
-          },
-        },
-        ballSpecs: [
-          { type: "ball", radius: 1, initialPosition: { x: 0, y: -5 } },
-        ],
-        coneSpecs: [
-          {
-            type: "cone",
-            baseColor: 0x980000,
-            height: 5,
-            radius: 0.5,
-            initialPosition: { x: 1.5, y: -25 },
-          },
-          {
-            type: "cone",
-            baseColor: 0x980000,
-            height: 5,
-            radius: 0.5,
-            initialPosition: { x: -1.5, y: -25 },
-          },
-          {
-            type: "cone",
-            baseColor: 0x980000,
-            height: 5,
-            radius: 0.5,
-            initialPosition: { x: 4.5, y: -25 },
-          },
-          {
-            type: "cone",
-            baseColor: 0x980000,
-            height: 5,
-            radius: 0.5,
-            initialPosition: { x: -4.5, y: -25 },
-          },
-          {
-            type: "cone",
-            baseColor: 0x980000,
-            height: 5,
-            radius: 0.5,
-            initialPosition: { x: 0, y: -22 },
-          },
-          {
-            type: "cone",
-            baseColor: 0x980000,
-            height: 5,
-            radius: 0.5,
-            initialPosition: { x: 3, y: -22 },
-          },
-          {
-            type: "cone",
-            baseColor: 0x980000,
-            height: 5,
-            radius: 0.5,
-            initialPosition: { x: -3, y: -22 },
-          },
-          {
-            type: "cone",
-            baseColor: 0x980000,
-            height: 5,
-            radius: 0.5,
-            initialPosition: { x: 1.5, y: -19 },
-          },
-          {
-            type: "cone",
-            baseColor: 0x980000,
-            height: 5,
-            radius: 0.5,
-            initialPosition: { x: -1.5, y: -19 },
-          },
-          {
-            type: "cone",
-            baseColor: 0x980000,
-            height: 5,
-            radius: 0.5,
-            initialPosition: { x: 0, y: -16 },
-          },
-        ],
-      };
-      setupArena(arenaConfig);
-    }
-
-    setupPlainArena();
     const setArena = (id: number) => {
       console.log("setArena id ", id);
+      let arenaConfig: ArenaConfig;
       switch (id) {
         case 1:
-          setupParkingLotArena();
+          arenaConfig = setupParkingLotArena();
           break;
         case 2:
-          setupZigZagArena();
+          arenaConfig = setupZigZagArena();
           break;
         case 3:
-          setupBowlingArena();
+          arenaConfig = setupBowlingArena();
           break;
         default:
-          setupPlainArena();
+          arenaConfig = setupPlainArena();
       }
+      setupArena(arenaConfig);
     };
-
     controls.linkSetArena(setArena);
-
-    return () => {
-      // clean up
-    };
+    setArena(0);
   }, [controls]);
 
   useEffect(() => {
